[Feat]2023.08.17 이벤트 리스트 내용 퍼블리싱
