+++ ssg_point_app_fe/app/(event)/endevent/page.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export default function Endevent() {
+  return (
+    <div>Endevent</div>
+  )
+}
