+++ ssg_point_app_fe/app/(event)/style.css
@@ -27,7 +27,7 @@
   width: 100%;
 }
 
-.event_list_box01 {
+.event_sort_box {
   height: 52px;
   /* display: -ms-flexbox; */
   display: flex;
@@ -37,7 +37,7 @@
   /* justify-content: flex-start; */
 }
 
-.event_list_box01 .select_box {
+.event_sort_box .select_box {
   width: 95px;
   height: 38px;
   border-radius: 0;
@@ -57,12 +57,39 @@ select option {
   position: relative;
   color: #000;
 }
-/* .event_list_box01 .select_box select {
-  height: 38px;
-  padding: 0 30px 0 10px;
-  font-size: 14px;
-  border: none;
+
+.event_list li a {
+  display: block;
   position: relative;
-  border-radius: 0;
-  color: #000;
-} */
+}
+.event_list li img{
+  width: 100%;
+}
+
+.event_list li .event_end_date {
+  padding: 20px 20px 35px;
+}
+
+.event_list li .event_end_date .event_tit {
+  font-size: 16px;
+  font-weight: 500;
+  line-height: 26px;
+  text-align: left;
+  white-space: nowrap;
+  overflow: hidden;
+  text-overflow: ellipsis;
+}
+
+.event_list li .end_date {
+  width: 82px;
+  height: 56px;
+  display: -ms-flexbox;
+  display: flex;
+  -ms-flex-pack: center;
+  justify-content: center;
+  position: absolute;
+  top: 10px;
+  right: 5px;
+  background: url(https://m.shinsegaepoint.com/img/event_end_date.fea1accb.png) no-repeat;
+  background-size: 100% auto;
+}
