+++ ssg_point_app_fe/app/(event)/winevent/page.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export default function Winevent() {
+  return (
+    <div>Winevent</div>
+  )
+}
