--- ssg_point_app_fe/app/event/[type]/page.tsx
@@ -1,11 +1,0 @@
-
-import React from 'react'
-
-export default function Event() {
-  
-
-  return (<>
-  </>
-  )
-}
-
