+++ ssg_point_app_fe/app/globals.css
@@ -49,7 +49,7 @@ body, ul, ol, li, h1, h2, h3, h4, h5, h6, p, button, input, textarea, figure, fi
 }
 
 .main_header {
-  position: fixed;
+  position: sticky;
   top: 0;
   left: 0;
   width: 100%;
@@ -57,6 +57,4 @@ body, ul, ol, li, h1, h2, h3, h4, h5, h6, p, button, input, textarea, figure, fi
   background-color: var(--ssg-point-secondary);
   box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
   padding: 1rem;
-  padding-bottom: 0;
-  position: static;
 }
