+++ ssg_point_app_fe/db.json
@@ -1,0 +1,46 @@
+{
+  "posts": [
+    {
+      "id": 1,
+      "title": "json-server",
+      "author": "typicode"
+    }
+  ],
+  "comments": [
+    {
+      "id": 1,
+      "body": "some comment",
+      "postId": 1
+    }
+  ],
+  "profile": {
+    "name": "typicode"
+  },
+  "event" : [
+    {
+      "id" : 1,
+      "event_id" : 30444,
+      "evnet_head" : "맘키즈클럽플러스 행사상품 안내(8월 2차)",
+      "reg_date" : "2023-08-12",
+      "evnet_start" : "2023-08-14",
+      "evnet_end" : "2023-09-11",
+      "event_thumbnail" : "https://mycloudmembership-prd.s3.ap-northeast-2.amazonaws.com/shinsegaepoint/public/svcm-fo/webapp/20230810/61f1891d-a549-48cf-bde7-f0daccb3ae3f.jpg",
+      "event_partner_id" : "200",
+      "event_type" : "단순연결",
+      "event_url" : "https://nextjs.org/docs/app/building-your-application/data-fetching/patterns"
+    },
+    {
+      "id" : 2,
+      "event_id" : 30456,
+      "evnet_head" : "현대카드 파리바게뜨팩&영화팩 구독료 지원 이벤트",
+      "reg_date" : "2023-08-12",
+      "evnet_start" : "2023-08-14",
+      "evnet_end" : "2023-09-11",
+      "event_thumbnail" : "https://mycloudmembership-prd.s3.ap-northeast-2.amazonaws.com/shinsegaepoint/public/svcm-fo/webapp/20230803/a2014220-2f8e-438a-a34b-4bd890ea63cd.png",
+      "event_partner_id" : "204",
+      "event_type" : "응모형",
+      "event_url" : "https://nextjs.org/docs/app/building-your-application/data-fetching/patterns"
+    }
+
+  ]
+}
